Refactors checklist modal into its own component
